v0.2.2: add recognition handlers
