+++ modules/input-laguages.mjs
@@ -1,0 +1,85 @@
+const languages = [
+  ['Afrikaans',       ['af-ZA']],
+  ['Bahasa Indonesia',['id-ID']],
+  ['Bahasa Melayu',   ['ms-MY']],
+  ['Català',          ['ca-ES']],
+  ['Čeština',         ['cs-CZ']],
+  ['Deutsch',         ['de-DE']],
+  ['English',         ['en-AU', 'Australia'],
+    ['en-CA', 'Canada'],
+    ['en-IN', 'India'],
+    ['en-NZ', 'New Zealand'],
+    ['en-ZA', 'South Africa'],
+    ['en-GB', 'United Kingdom'],
+    ['en-US', 'United States']],
+  ['Español',         ['es-AR', 'Argentina'],
+    ['es-BO', 'Bolivia'],
+    ['es-CL', 'Chile'],
+    ['es-CO', 'Colombia'],
+    ['es-CR', 'Costa Rica'],
+    ['es-EC', 'Ecuador'],
+    ['es-SV', 'El Salvador'],
+    ['es-ES', 'España'],
+    ['es-US', 'Estados Unidos'],
+    ['es-GT', 'Guatemala'],
+    ['es-HN', 'Honduras'],
+    ['es-MX', 'México'],
+    ['es-NI', 'Nicaragua'],
+    ['es-PA', 'Panamá'],
+    ['es-PY', 'Paraguay'],
+    ['es-PE', 'Perú'],
+    ['es-PR', 'Puerto Rico'],
+    ['es-DO', 'República Dominicana'],
+    ['es-UY', 'Uruguay'],
+    ['es-VE', 'Venezuela']],
+  ['Euskara',         ['eu-ES']],
+  ['Français',        ['fr-FR']],
+  ['Galego',          ['gl-ES']],
+  ['Hrvatski',        ['hr_HR']],
+  ['IsiZulu',         ['zu-ZA']],
+  ['Íslenska',        ['is-IS']],
+  ['Italiano',        ['it-IT', 'Italia'],
+    ['it-CH', 'Svizzera']],
+  ['Magyar',          ['hu-HU']],
+  ['Nederlands',      ['nl-NL']],
+  ['Norsk bokmål',    ['nb-NO']],
+  ['Polski',          ['pl-PL']],
+  ['Português',       ['pt-BR', 'Brasil'],
+    ['pt-PT', 'Portugal']],
+  ['Română',          ['ro-RO']],
+  ['Slovenčina',      ['sk-SK']],
+  ['Suomi',           ['fi-FI']],
+  ['Svenska',         ['sv-SE']],
+  ['Türkçe',          ['tr-TR']],
+  ['български',       ['bg-BG']],
+  ['Pусский',         ['ru-RU']],
+  ['Српски',          ['sr-RS']],
+  ['한국어',            ['ko-KR']],
+  ['中文',             ['cmn-Hans-CN', '普通话 (中国大陆)'],
+    ['cmn-Hans-HK', '普通话 (香港)'],
+    ['cmn-Hant-TW', '中文 (台灣)'],
+    ['yue-Hant-HK', '粵語 (香港)']],
+  ['日本語',           ['ja-JP']],
+  ['Lingua latīna',   ['la']]
+];
+
+const populateWithLanguageOptions = (selectElement) => {
+  for (let i = 0; i < languages.length; i++) {
+    selectElement.options[i] = new Option(languages[i][0], `${i}`);
+  }
+}
+
+const createLanguageHandler = ( languageSelectElement, dialectSelectElement ) => {
+  return function (languageIndex = 0, dialectIndex = 0) {
+    languageSelectElement.selectedIndex = languageIndex;
+    dialectSelectElement.length = 0;
+    const dialects = languages[languageSelectElement.selectedIndex];
+    for (let i = 1; i < dialects.length; i++) {
+      dialectSelectElement.options.add(new Option(dialects[i][1], dialects[i][0]));
+    }
+    dialectSelectElement.style.visibility = dialects[1].length === 1 ? 'hidden' : 'visible';
+    dialectSelectElement.selectedIndex = dialectIndex;
+  };
+};
+
+export { populateWithLanguageOptions, createLanguageHandler }
